requirements.txt
pdf2image
pillow
pytest
pytesseract
pandas
openpyxl
pdfplumber
